shard manager bug fixes + containerized, server dependencies issue fixed, make file cleaned, permission issues WIP
