+++ Assignment3/Makefile
@@ -13,10 +13,4 @@ clean:
 	docker image rm ds_server:latest
 
 restart:
-	docker compose down --rmi all
-	docker stop $$(docker ps -a -q)
-	docker rm $$(docker ps -a -q)
-	docker network rm pub
-	docker image rm ds_server:latest
-	docker build -t ds_server:latest ./server --platform linux/x86_64
-	docker compose up -d
+	make clean ; make all
