+++ Assignment3/shard_manager/requirements.txt
@@ -1,0 +1,2 @@
+Flask==2.3.2
+Requests==2.31.0
